+++ package/src/base/rowid_with_config.rs
@@ -2,7 +2,7 @@ use std::io;
 use std::time::SystemTime;
 
 use crate::common::configs::{CHAR_LIST, RANDOMNESS_LENGTH};
-use crate::common::errors::RWC_DONE_CHAR_LIST_LENGTH_ERR;
+use crate::common::errors::RowIDError;
 use crate::functions::decode::{decode as _decode, DecodeOptions};
 use crate::functions::encode::{encode as _encode, EncodeOptions};
 use crate::functions::generate::{
@@ -17,9 +17,10 @@ use crate::functions::verify::{
 };
 
 /// This struct contains the state of the `RowIDWithConfig` struct.
-pub struct RowIDWithConfigState {
+#[derive(Debug, Clone, Copy)]
+pub struct RowIDWithConfigState<'a> {
     /// The list of characters used in the current function.
-    pub char_list: String,
+    pub char_list: &'a str,
     /// The length of randomness used in the current function.
     pub randomness_length: usize,
 }
@@ -35,12 +36,13 @@ pub struct RowIDWithConfigState {
 /// let rwc: RowIDWithConfigResult =
 ///     RowIDWithConfig::new().done().unwrap();
 /// ```
-pub struct RowIDWithConfigResult {
+#[derive(Debug, Clone, Copy)]
+pub struct RowIDWithConfigResult<'a> {
     /// Represents function's customization configurations.
-    pub state: RowIDWithConfigState,
+    pub state: RowIDWithConfigState<'a>,
 }
 
-impl RowIDWithConfigResult {
+impl<'a> RowIDWithConfigResult<'a> {
     /// This function generates a unique ID
     /// that is almost impossible to duplicate.
     ///
@@ -55,7 +57,7 @@ impl RowIDWithConfigResult {
     /// ```
     pub fn rowid(&self) -> String {
         _rowid(RowIDOptions {
-            char_list: self.state.char_list.clone(),
+            char_list: self.state.char_list,
             randomness_length: self.state.randomness_length,
         })
     }
@@ -77,10 +79,7 @@ impl RowIDWithConfigResult {
         &self,
         system_time: SystemTime,
     ) -> io::Result<String> {
-        _encode(EncodeOptions {
-            char_list: self.state.char_list.clone(),
-            system_time,
-        })
+        _encode(EncodeOptions { char_list: self.state.char_list, system_time })
     }
 
     /// This function decodes the ID into a timestamp in milliseconds.
@@ -93,17 +92,13 @@ impl RowIDWithConfigResult {
     ///
     /// let rwc: RowIDWithConfigResult =
     ///     RowIDWithConfig::new().done().unwrap();
-    /// let id: String = "ABC123".to_string();
-    /// let decoded: SystemTime = rwc.decode(id).unwrap();
+    /// let decoded: SystemTime = rwc.decode("ABC123").unwrap();
     /// ```
     pub fn decode(
         &self,
-        encoded: String,
+        encoded: &str,
     ) -> io::Result<SystemTime> {
-        _decode(DecodeOptions {
-            char_list: self.state.char_list.clone(),
-            encoded,
-        })
+        _decode(DecodeOptions { char_list: self.state.char_list, encoded })
     }
 
     /// This function generates an ID based on the input.
@@ -125,7 +120,7 @@ impl RowIDWithConfigResult {
         randomness_length: Option<usize>,
     ) -> GenerateResult {
         _generate(GenerateOptions {
-            char_list: self.state.char_list.clone(),
+            char_list: self.state.char_list,
             system_time,
             randomness_length: match randomness_length {
                 | Some(l) => l,
@@ -143,17 +138,13 @@ impl RowIDWithConfigResult {
     ///
     /// let rwc: RowIDWithConfigResult =
     ///     RowIDWithConfig::new().done().unwrap();
-    /// let id: String = "ABC123".to_string();
-    /// let result: VerifyResult = rwc.verify(id);
+    /// let result: VerifyResult = rwc.verify("ABC123");
     /// ```
     pub fn verify(
         &self,
-        encoded: String,
+        encoded: &str,
     ) -> VerifyResult {
-        _verify(VerifyOptions {
-            char_list: self.state.char_list.clone(),
-            encoded,
-        })
+        _verify(VerifyOptions { char_list: self.state.char_list, encoded })
     }
 
     /// This function generates randomness.
@@ -172,7 +163,7 @@ impl RowIDWithConfigResult {
         randomness_length: usize,
     ) -> String {
         _get_randomness(GetRandomnessOptions {
-            char_list: self.state.char_list.clone(),
+            char_list: self.state.char_list,
             randomness_length,
         })
     }
@@ -191,11 +182,12 @@ impl RowIDWithConfigResult {
 /// let rwc: RowIDWithConfigResult =
 ///     RowIDWithConfig::new().done().unwrap();
 /// ```
-pub struct RowIDWithConfig {
-    state: RowIDWithConfigState,
+#[derive(Debug, Clone, Copy)]
+pub struct RowIDWithConfig<'a> {
+    state: RowIDWithConfigState<'a>,
 }
 
-impl RowIDWithConfig {
+impl<'a> RowIDWithConfig<'a> {
     /// Creates a new `RowIDWithConfig` with default values.
     ///
     /// ## Example
@@ -209,14 +201,14 @@ impl RowIDWithConfig {
     pub fn new() -> Self {
         Self {
             state: RowIDWithConfigState {
-                char_list: CHAR_LIST.to_string(),
+                char_list: CHAR_LIST,
                 randomness_length: RANDOMNESS_LENGTH,
             },
         }
     }
 
     /// The list of characters that can be used in the RowID,
-    /// it must be longer or equal to `28`.
+    /// The length of it must be greater or equal to `28`.
     ///
     /// default: `0123456789ABCDEFGHJKMNPQRSTVWXYZ`
     ///
@@ -226,7 +218,7 @@ impl RowIDWithConfig {
     /// use rowid::{RowIDWithConfig, RowIDWithConfigResult};
     ///
     /// let rwc: RowIDWithConfigResult = RowIDWithConfig::new()
-    ///     .char_list("0123456789ABCDEFGHJKMNPQRSTVWXYZ".to_string())
+    ///     .char_list("0123456789ABCDEFGHJKMNPQRSTVWXYZ")
     ///     .done()
     ///     .unwrap();
     ///
@@ -234,7 +226,7 @@ impl RowIDWithConfig {
     /// ```
     pub fn char_list(
         mut self,
-        list: String,
+        list: &'a str,
     ) -> Self {
         self.state.char_list = list;
         self
@@ -278,24 +270,24 @@ impl RowIDWithConfig {
     ///     RowIDWithConfig::new().done().unwrap();
     /// let id: String = rwc.rowid();
     /// ```
-    pub fn done(self) -> io::Result<RowIDWithConfigResult> {
+    pub fn done(self) -> io::Result<RowIDWithConfigResult<'a>> {
         if self.state.char_list.len() < 28 {
             return Err(io::Error::new(
                 io::ErrorKind::InvalidInput,
-                RWC_DONE_CHAR_LIST_LENGTH_ERR,
+                RowIDError::CharListLength.as_str(),
             ));
         }
 
         Ok(RowIDWithConfigResult {
             state: RowIDWithConfigState {
-                char_list: self.state.char_list.clone(),
+                char_list: self.state.char_list,
                 randomness_length: self.state.randomness_length,
             },
         })
     }
 }
 
-impl Default for RowIDWithConfig {
+impl<'a> Default for RowIDWithConfig<'a> {
     fn default() -> Self {
         Self::new()
     }