+++ package/CHANGELOG.md
@@ -1,3 +1,23 @@
+## 0.2.0 (2024-10-13)
+
+### Breaking Changes
+
+- Changes in accepted value type of `decode`:
+    - `String` => `&str`
+- Changes in accepted value type of `verify`:
+    - `String` => `&str`
+- Changes in accepted value type of `char_list` in `RowIDWithConfig`:
+    - `String` => `&str`
+- Merge different error messages into `RowIDError` enum
+
+### What's New
+
+- Add different derives for different structs
+
+### What's Changed
+
+- Updates in documentation
+
 ## 0.1.1 (2024-08-04)
 
 ### What's Changed